created user-filtering component
