+++ app/templates/components/popup-modal.hbs
@@ -13,7 +13,7 @@
        {{/each}}   --}}
 
         {{#each popup-btns as |btn|}}
-        {{#link-to 'users' class="popup-link" invokeAction=(if (not-eq btn.actionName 'null') (action (action 'buttonAction' btn.actionName))) 
+        {{#link-to path class="popup-link" invokeAction=(if (not-eq btn.actionName 'null') (action (action 'buttonAction' btn.actionName))) 
         preventDefault=false}}<BsButton class="popup-buttons">
           
         {{btn.title}}</BsButton>{{/link-to}}
